feat(backend now sends result to frontend) + fix(positioning the 'and' when formating the final word)

+++ NumbersToWords/NumToWord.cs
@@ -3,7 +3,7 @@
     public class NumToWord
     {
         public int Id { get; set; }
-        public int NumInput { get; set; }
+        public float NumInput { get; set; }
         public string? NumConvertedOutput { get; set; }
     }
 }
